Make display event its own file
