+++ cli.js
@@ -1,12 +1,11 @@
 const puppeteer = require('puppeteer');
-const terminalLink = require('terminal-link');
 const chalk = require('chalk');
 const ora = require('ora');
 const clear = require('clear');
 const inquirer = require('inquirer');
 const fs = require('fs');
 
-const dateAssesor = require('./src/dateAssesor');
+const displayEvents = require('./src/displayEvents');
 
 const parallel = 4;
 const strings = require('./src/strings.json');
@@ -31,7 +30,6 @@ const intializeApp = () => {
 		});
 	}
 };
-
 
 const saveTrackedPages = () => {
 	let formattedPages = pages.toString().replace(/\n/g, '');
@@ -227,50 +225,6 @@ const crawlFacebook = async (pages, parallel) => {
 	}
 
 	console.log('\n' + strings.english.success.crawlingCompleted);
-};
-
-const displayEvents = (events) => {
-	try {
-		for (let i = 0; i < events.recurringEvents.descriptions.length; i++) {
-			let j = i + 1;
-
-			console.log(chalk.underline('Title:') + ' ' + events.recurringEvents.titles[i]);
-			console.log(chalk.underline('Description:') + ' ' + events.recurringEvents.descriptions[i]);
-
-			if (i === 0) {
-				console.log(chalk.underline('Date:') + ' ' + events.recurringEvents.dates[i] + ' & ' + events.recurringEvents.dates[j]);
-			} else {
-				console.log(chalk.underline('Date:') + ' ' + events.recurringEvents.dates[i * 2] + ' & ' + events.recurringEvents.dates[i * 2 + 1]);
-			}
-			console.log(
-				chalk.blue(
-					chalk.underline(
-						terminalLink('Link to the event', 'https://facebook.com' + events.recurringEvents.link[i])
-					)
-				)
-			);
-			console.log('\n');
-		}
-
-		for (let i = 0; i < events.pastAndUpcoming.titles.length; i++) {
-			const shouldEventBeDisplayed = dateAssesor(events.pastAndUpcoming.dates[i]);
-
-			if (shouldEventBeDisplayed) {
-				console.log(chalk.underline('Title:') + ' ' + events.pastAndUpcoming.titles[i]);
-				console.log(chalk.underline('Date:') + ' ' + events.pastAndUpcoming.dates[i]);
-				console.log(
-					chalk.blue(
-						chalk.underline(
-							terminalLink('Link to the event', 'https://facebook.com' + events.pastAndUpcoming.link[i])
-						)
-					)
-				);
-				console.log('\n');
-			}
-		}
-	} catch (err) {
-		console.log(err);
-	}
 };
 
 intializeApp();
